maybe oscam is not installed let's avoid commands

+++ src/SoftcamManager.py
@@ -534,7 +534,7 @@ class SoftcamAutoPoller:
 	def JobStart(self):
 		self.autostartcams = config.softcammanager.softcams_autostart.value
 		self.Console = Console()
-		if not path.exists('/usr/softcams/oscam'):
+		if not path.exists('/usr/softcams/oscam') and path.exists('/usr/bin/oscam'):
 			self.Console.ePopen('ln -s /usr/bin/*oscam* /usr/softcams/')
 		if not path.exists('/usr/softcams/ncam') and path.exists('/usr/bin/ncam'):
 		    self.Console.ePopen('ln -s /usr/bin/ncam /usr/softcams/')
@@ -636,7 +636,7 @@ class SoftcamAutoPoller:
 						if softcamcheck.lower().startswith('oscam') or softcamcheck.lower().startswith('ncam'):
 							self.Console.ePopen("ps.procps | grep softcams | grep -v grep | awk 'NR==1' | awk '{print $5}'| awk  -F'[/]' '{print $4}' > /tmp/softcamRuningCheck.tmp")
 							sleep(2)
-							
+
 #							open('/tmp/softcamRuningCheck.tmp')
 							self.Console.ePopen('ulimit -s 1024;/usr/softcams/' + softcamcheck + " -b")
 							sleep(10)
